Vukasin IT26-2019 Projekat

+++ Vukasin Stanisic IT26-2019/src/Drawing/FrmDrawing.java
@@ -4,6 +4,7 @@ import java.awt.BorderLayout;
 import java.awt.EventQueue;
 
 import javax.swing.JFrame;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.border.EmptyBorder;
 
@@ -148,10 +149,16 @@ public class FrmDrawing extends JFrame {
 	protected void delete(){
 		Shape selected=pnlCenter.getSelected();
 		if(selected!=null){
-			pnlCenter.getShapes().remove(selected);
-			pnlCenter.deleteSelected();
+			int reply=JOptionPane.showConfirmDialog(null,"Are you sure you want to delete this?","Delete?",JOptionPane.YES_NO_OPTION);
+			if(reply==JOptionPane.YES_OPTION)
+			{
+				pnlCenter.getShapes().remove(selected);
+				pnlCenter.deleteSelected();
+				pnlCenter.repaint();
+			}
+			
 		}
-			pnlCenter.repaint();
+			
 		
 	}
 	protected void modify() throws Exception{
